Data feed for 2013 conference
With a lot of dirty dirty hacking I was able to get the 2013 data feed
being processed.
I took the hacky last year code, and put it through a meat grinder to
make hack it up even more while splicing in a XML parser for this year's
feed.
I REALLY need to rewrite this entire thing from scratch for next year.